alter table harga_perTon bigInt

+++ farming_v3/models.py
@@ -82,7 +82,7 @@ class Tanaman(models.Model):
     nama_tanaman = models.CharField(max_length=100, unique=True, db_index=True)
     jenis = models.CharField(max_length=100)
     waktu_tanam_hari = models.IntegerField(default=0)
-    harga_perTon = models.IntegerField(default=0)
+    harga_perTon = models.BigIntegerField(default=0)
     peluang_hama = models.ForeignKey(Hama, on_delete=models.CASCADE)
     owner = models.ForeignKey(User, related_name='tanaman', on_delete=models.CASCADE)
     deskripsi = models.TextField(default="deskripsi tanaman")
